Fusion: move Future to separate file
