+++ pykokkos/core/fusion/trace.py
@@ -4,48 +4,8 @@ from types import NoneType
 from typing import Any, Callable, Dict, List, Optional, Set, Union
 
 from pykokkos.interface import ExecutionPolicy, ViewType
-from pykokkos.runtime import runtime_singleton
 
-
-class Future:
-    def __init__(self) -> None:
-        self.value = None
-
-    def assign_value(self, value) -> None:
-        self.value = value
-
-    def __add__(self, other):
-        self.flush_trace()
-        return self.value + other
-
-    def __sub__(self, other):
-        self.flush_trace()
-        return self.value - other
-
-    def __mul__(self, other):
-        self.flush_trace()
-        return self.value * other
-
-    def __truediv__(self, other):
-        self.flush_trace()
-        return self.value / other
-
-    def __floordiv__(self, other):
-        self.flush_trace()
-        return self.value // other
-
-    def __str__(self):
-        self.flush_trace()
-        return str(self.value)
-
-    def __repr__(self) -> str:
-        return str(f"Future(value={self.value})")
-
-    def flush_trace(self) -> None:
-        if self.value is not None:
-            return
-        runtime_singleton.runtime.flush_data(self)
-        assert self.value is not None
+from .future import Future
 
 
 @dataclass
